// apps/serverless-admin/src/lib/schemas.ts
export const logLineSchema = z.object({
  sequence: z.number(),
  accountId: z.string().optional(),
  namespace: z.string().optional(),
  functionName: z.string().optional(),
  functionVersion: z.string().optional(),
  requestId: z.string().optional(),
  workerId: z.string().optional(),
  sandboxId: z.string().optional(),
  stream: z.string(),
  level: z.string().optional(),
  message: z.string(),
  timestamp: z.string(),
})
// ...
export type LogLine = z.infer<typeof logLineSchema>